refactor(slack-bot): extract repo-status.js from index.js (behavior-preserving)
